Refactor: 충전하기 Button으로 변경
