+++ src/pages/List.module.scss
@@ -73,14 +73,14 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
-  padding: 20px;
+  padding: 36px 64px;
   border: 1px solid rgba($color: #f1eef9, $alpha: 0.8);
   background-color: black;
   color: white;
   border-radius: 8px;
 
-  @media screen and (min-width: $tablet) {
-    padding: 36px 64px;
+  @media screen and (max-width: $tablet) {
+    padding: 20px;
   }
 
   .leftContent {
@@ -89,13 +89,13 @@
     flex-direction: column;
 
     & > span {
-      font-size: 1.2rem;
+      font-size: 1.6rem;
       font-weight: 400;
       color: rgba($color: white, $alpha: 0.6);
       user-select: none;
 
-      @media screen and (min-width: $tablet) {
-        font-size: 1.6rem;
+      @media screen and (max-width: $tablet) {
+        font-size: 1.2rem;
       }
     }
 
@@ -104,34 +104,26 @@
       align-items: center;
 
       span {
-        font-size: 2rem;
+        font-size: 2.4rem;
         font-weight: 700;
         line-height: 26px;
         user-select: none;
 
-        @media screen and (min-width: $tablet) {
-          font-size: 2.4rem;
+        @media screen and (max-width: $tablet) {
+          font-size: 2rem;
         }
       }
     }
   }
 
   .rightContent {
-    font-size: 1.3rem;
+    font-size: 1.6rem;
     font-weight: 700;
     line-height: 26px;
-    color: var(--brand-orange);
     letter-spacing: 0.5px;
-    cursor: pointer;
-    user-select: none;
 
-    @media screen and (min-width: $tablet) {
-      font-size: 1.6rem;
-    }
-
-    &:focus,
-    &:hover {
-      color: var(--brand-pink);
+    @media screen and (max-width: $tablet) {
+      font-size: 1.3rem;
     }
   }
 }
